requirements.txt
dj-static==0.0.6
Django==3.1.4
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.5
django-allauth==0.44.0
django-bootstrap-breadcrumbs==0.9.2
django-bootstrap3==14.2.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-countries==5.3.3
django-crispy-forms==1.11.0
django-currencies==0.10.1
django-debug-toolbar==3.1.1
django-environ==0.4.5
django-extensions==3.1.0
django-filter==2.4.0
django-guardian==2.1.0
django-honeypot==1.0.1
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-material==1.9.0
django-model-utils==3.2.0
django-mptt==0.11.0
django-notifications-hq==1.5.0
django-phonenumber-field==2.0.0
django-restframework==0.0.1
django-searchable-select==1.5.0
django-social-share==2.2.1
django-storages==1.9.1
django-summernote==0.8.11.6
django-tables2==2.4.0
django-taggit==1.3.0
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docopt==0.6.2
drf-yasg==1.20.0
Faker==4.16.0
graphviz==0.16
gunicorn==20.1.0
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
isort==5.6.4
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonfield==2.1.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
mysqlclient==2.0.3
numpy==1.20.3
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
pep8==1.7.1
phonenumberslite==8.12.24
Pillow==8.0.1

